Clean code and comments
