+++ nbproject/private/private.xml
@@ -3,7 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/home/alex/NetBeansProjects/ClienteRubyWebService2/src/clienterubywebservice/MainInterface.java</file>
+            <file>file:/media/sf_Projetos/tec-integracion-soluciones/github-webservice/java-ws-client/src/clientDefaultPackage/ClientInterface.java</file>
         </group>
     </open-files>
 </project-private>
