+++ nbproject/project.properties
@@ -56,7 +56,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=clienterubywebservice.ClienteRubyWebService
+main.class=clientDefaultPackage.RunClientInterface
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
