+++ src/clientDefaultPackage/RunClientInterface.java
@@ -1,0 +1,13 @@
+package clientDefaultPackage;
+
+/**
+ * @author Alex y Caio
+ */
+public class RunClientInterface {
+
+    // Hace una instacia de la interfaz, y la mostra en la pantalla
+    public static void main(String[] args) {
+        ClientInterface gui = new ClientInterface();
+        gui.setVisible(true);
+    }    
+}
