--- src/clienterubywebservice/ClienteRubyWebService.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package clienterubywebservice;
-
-/**
- *
- * @author alex
- */
-public class ClienteRubyWebService {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        // TODO code application logic here
-        MainInterface gui = new MainInterface();
-
-        gui.setVisible(true);
-    }
-    
-}
