--- src/clienterubywebservice/MainInterface.java
@@ -1,271 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package clienterubywebservice;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.*;
-
-/**
- *
- * @author alex
- */
-public class MainInterface extends javax.swing.JFrame {
-
-    /**
-     * Creates new form mainInterface
-     */
-    public MainInterface() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jLabelTitulo = new javax.swing.JLabel();
-        textNumero1 = new javax.swing.JTextField();
-        textNumero2 = new javax.swing.JTextField();
-        buttonEnviar = new javax.swing.JButton();
-        jLabelRespuesta = new javax.swing.JLabel();
-        comboBoxOperaciones = new javax.swing.JComboBox<>();
-        jLabelOperacion = new javax.swing.JLabel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setTitle("Cliente Ruby");
-        setLocation(new java.awt.Point(0, 0));
-
-        jLabelTitulo.setFont(new java.awt.Font("Noto Sans", 1, 12)); // NOI18N
-        jLabelTitulo.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabelTitulo.setText("Calculadora");
-
-        textNumero1.setHorizontalAlignment(javax.swing.JTextField.CENTER);
-
-        textNumero2.setHorizontalAlignment(javax.swing.JTextField.CENTER);
-
-        buttonEnviar.setText("Enviar");
-        buttonEnviar.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                buttonEnviarActionPerformed(evt);
-            }
-        });
-
-        jLabelRespuesta.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-
-        comboBoxOperaciones.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Suma", "Resta", "Multiplicación", "División", "Potencia", "Raíz" }));
-        comboBoxOperaciones.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                comboBoxOperacionesActionPerformed(evt);
-            }
-        });
-
-        jLabelOperacion.setText("+");
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jLabelRespuesta, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addContainerGap()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabelTitulo, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addGroup(layout.createSequentialGroup()
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addGroup(layout.createSequentialGroup()
-                                        .addGap(12, 12, 12)
-                                        .addComponent(comboBoxOperaciones, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                                    .addGroup(layout.createSequentialGroup()
-                                        .addComponent(textNumero1, javax.swing.GroupLayout.PREFERRED_SIZE, 55, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                        .addComponent(jLabelOperacion)
-                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                        .addComponent(textNumero2, javax.swing.GroupLayout.PREFERRED_SIZE, 55, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                                .addGap(0, 0, Short.MAX_VALUE))))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(57, 57, 57)
-                        .addComponent(buttonEnviar)))
-                .addContainerGap())
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jLabelTitulo)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(textNumero1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(textNumero2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabelOperacion))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(comboBoxOperaciones, javax.swing.GroupLayout.PREFERRED_SIZE, 24, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(buttonEnviar)
-                .addGap(18, 18, 18)
-                .addComponent(jLabelRespuesta, javax.swing.GroupLayout.PREFERRED_SIZE, 23, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void buttonEnviarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_buttonEnviarActionPerformed
-        String a, b;
-        int type = comboBoxOperaciones.getSelectedIndex();
-        
-        a = textNumero1.getText();
-        b = textNumero2.getText();
-        
-        // Checa si los campos contienen números, si contienen hace la petición 
-        if ((!isNumeric(a) && type == 5) || (!isNumeric(a) || (!isNumeric(b) && type != 5))) {
-            jLabelRespuesta.setText(type == 5 ? "Digite un número!" : "Digite los números!");
-        } else {
-            String operacion = "";
-            
-            // Determina la operación a ser solicitada, según selecionado por el usuario
-            switch (type) {
-                case 0:
-                    operacion = "suma";
-                    break;
-                case 1:
-                    operacion = "resta";
-                    break;
-                case 2:
-                    operacion = "multiplicacion";
-                    break;
-                case 3:
-                    operacion = "division";
-                    break;
-                case 4:
-                    operacion = "potencia";
-                    break;
-                case 5:
-                    operacion = "raiz";
-                    break;
-            }             
-            
-            try {
-                // Crea una URL usando los parámetros obtenidos
-                URL url = new URL("http://localhost:4567/" + operacion + "?a=" + a + "&b=" + b);
-
-                // Hace una petición HTTP para el URL indicado
-                HttpURLConnection con = (HttpURLConnection) url.openConnection();
-                
-                // Interpreta la respuesta del servidor
-                BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));  
-                
-                // Escribe la respuesta en la pantalla, para el usuario
-                jLabelRespuesta.setText(in.readLine());
-
-                // Cierra el buffer y la conexión
-                in.close();
-                con.disconnect();
-            } catch (IOException ex) {
-                // Informa el usuario que ocurrio un error, y indica el error en la consola
-                jLabelRespuesta.setText("Ocurrio un error!");
-                ex.printStackTrace();
-            }
-        }
-    }//GEN-LAST:event_buttonEnviarActionPerformed
-
-    private void comboBoxOperacionesActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_comboBoxOperacionesActionPerformed
-        // TODO add your handling code here:
-        int type = comboBoxOperaciones.getSelectedIndex();
-        
-        textNumero2.setEditable(true);
-        
-        switch (type) {
-            case 0:
-                jLabelOperacion.setText("+");
-                break;
-            case 1:
-                jLabelOperacion.setText("-");
-                break;
-            case 2:
-                jLabelOperacion.setText("*");
-                break;
-            case 3:
-                jLabelOperacion.setText("/");
-                break;
-            case 4:
-                jLabelOperacion.setText("^");
-                break;
-            case 5:
-                jLabelOperacion.setText("√");
-                textNumero2.setText("");
-                textNumero2.setEditable(false);
-                break;
-        }
-    }//GEN-LAST:event_comboBoxOperacionesActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(MainInterface.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(MainInterface.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(MainInterface.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(MainInterface.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new MainInterface().setVisible(true);
-            }
-        });
-    }
-    
-    /* Método creado por el usuario "CraigTP". 
-     * Disponible en el link: https://stackoverflow.com/questions/1102891/how-to-check-if-a-string-is-numeric-in-java */
-    public static boolean isNumeric(String str) {  
-        try {  
-            double d = Double.parseDouble(str);  
-        } catch (NumberFormatException nfe) {  
-            return false;  
-        }
-        
-        return true;  
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton buttonEnviar;
-    private javax.swing.JComboBox<String> comboBoxOperaciones;
-    private javax.swing.JLabel jLabelOperacion;
-    private javax.swing.JLabel jLabelRespuesta;
-    private javax.swing.JLabel jLabelTitulo;
-    private javax.swing.JTextField textNumero1;
-    private javax.swing.JTextField textNumero2;
-    // End of variables declaration//GEN-END:variables
-}
